Move ConsoleStyle to new Console namespace
